refactor(Repeater): Split ViewportManager and port Downlevel variant
Reorganize ViewportManager hierarchy to match WinUI source layout (commit 5f9e85113):
- ViewportManager.cs: abstract base with member order matching WinUI h (MadeAnchor at end); add ResetLayoutRealizationWindowCacheBuffer abstract
- ViewportManagerWithPlatformFeatures split into .cs (decl), .h.mux.cs (fields + inline GetLayoutExtent/GetOrigin/MadeAnchor/HasScroller), .mux.cs (bodies in cpp source order), .uno.cs (native-only viewport cache)
- ViewportManagerDownlevel: new port of the IRepeaterScrollingSurface-based viewport manager, split into .cs/.h.mux.cs/.mux.cs per winui-port
