+++ src/Uno.UI/UI/Xaml/Controls/Repeater/ViewportManager.cs
@@ -1,46 +1,47 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
-// ItemsRepeater.cpp, commit 1cf9f1c
+// MUX Reference ViewportManager.h, commit 5f9e85113
 
 using Windows.Foundation;
 using Microsoft.UI.Xaml;
 
-namespace Microsoft.UI.Xaml.Controls
+namespace Microsoft.UI.Xaml.Controls;
+
+internal abstract partial class ViewportManager
 {
-	internal abstract partial class ViewportManager
-	{
-		public abstract UIElement SuggestedAnchor { get; }
+	public abstract UIElement SuggestedAnchor { get; }
 
-		public abstract UIElement MadeAnchor { get; }
+	public abstract double HorizontalCacheLength { get; set; }
 
-		public abstract double HorizontalCacheLength { get; set; }
+	public abstract double VerticalCacheLength { get; set; }
 
-		public abstract double VerticalCacheLength { get; set; }
+	public abstract Rect GetLayoutVisibleWindow();
 
-		public abstract Rect GetLayoutVisibleWindow();
+	public abstract Rect GetLayoutRealizationWindow();
 
-		public abstract Rect GetLayoutRealizationWindow();
+	public abstract void SetLayoutExtent(Rect layoutExtent);
 
-		public abstract void SetLayoutExtent(Rect extent);
+	public abstract Rect GetLayoutExtent();
 
-		public abstract Rect GetLayoutExtent();
+	public abstract Point GetOrigin();
 
-		public abstract Point GetOrigin();
+	public abstract void OnLayoutChanged(bool isVirtualizing);
 
-		public abstract void OnLayoutChanged(bool isVirtualizing);
+	public abstract void OnElementPrepared(UIElement element);
 
-		public abstract void OnElementPrepared(UIElement element);
+	public abstract void OnElementCleared(UIElement element);
 
-		public abstract void OnElementCleared(UIElement element);
+	public abstract void OnOwnerMeasuring();
 
-		public abstract void OnOwnerMeasuring();
+	public abstract void OnOwnerArranged();
 
-		public abstract void OnOwnerArranged();
+	public abstract void OnMakeAnchor(UIElement anchor, bool isAnchorOutsideRealizedRange);
 
-		public abstract void OnMakeAnchor(UIElement anchor, bool isAnchorOutsideRealizedRange);
+	public abstract void OnBringIntoViewRequested(BringIntoViewRequestedEventArgs args);
 
-		public abstract void OnBringIntoViewRequested(BringIntoViewRequestedEventArgs args);
+	public abstract void ResetLayoutRealizationWindowCacheBuffer();
 
-		public abstract void ResetScrollers();
-	}
+	public abstract void ResetScrollers();
+
+	public abstract UIElement MadeAnchor { get; }
 }
